--- packages/bsc-cross-sdk/src/MultiMessage.ts
@@ -1,1 +1,0 @@
-export default class MultiMessage {}
